Rename filenames.
