+++ backendV2/src/models/houseModel.ts
